requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
fastapi==0.115.2
graphene==3.3
graphql-core==3.2.5
graphql-relay==3.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
starlette-graphene3==0.6.0
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
